+++ src/App.svelte
@@ -2,14 +2,16 @@
   import { Router, Route } from "svelte-routing"
   import { siteName } from "./js/constants"
 
-  // COMPONENTS
-  import Header from "./components/Header.svelte"
-  import Footer from "./components/Footer.svelte"
-
   // ROUTES/PAGES
   import About from "./routes/About.svelte"
-  import Contact from "./routes/Contact.svelte"
   import Blog from "./routes/Blog.svelte"
+  import Contact from "./routes/Contact.svelte"
+  import FullPage from "./routes/FullPage.svelte"
+
+  import Test from "./routes/Test.svelte"
+
+  // COMPONENTS
+  import ColorModeLogic from "./components/ColorModeLogic.svelte"
 
   // PROPS
   export let url = "" // svelte-routing requirement
@@ -18,22 +20,18 @@
 <svelte:head>
   <title>{siteName}</title>
 </svelte:head>
+
 <Router {url}>
   <page>
-    <header>
-      <Header />
-    </header>
-    <main>
-      <Route path="/" component={About} />
-      <Route path="about" component={About} />
-      <Route path="contact" component={Contact} />
-      <Route path="blog/*" component={Blog} />
-    </main>
-    <footer>
-      <Footer />
-    </footer>
+    <Route path="/" component={About} />
+    <Route path="about" component={About} />
+    <Route path="blog/*" component={Blog} />
+    <Route path="contact" component={Contact} />
+    <Route path="fullpage" component={FullPage} />
+    <Route path="test" component={Test} />
   </page>
 </Router>
+<ColorModeLogic />
 
 <style>
   page {
@@ -43,30 +41,5 @@
     width: 100vw;
     position: relative;
     background-color: var(--color_body);
-  }
-
-  header,
-  footer {
-    flex: 0 1;
-    display: flex;
-    padding: 10px 15px;
-    background-color: var(--color_bg_1);
-  }
-  header {
-    justify-content: space-between;
-    border-bottom: 1px solid var(--color_border_1);
-  }
-  footer {
-    border-top: 1px solid var(--color_border_1);
-    flex-direction: column;
-    justify-content: center;
-    align-items: center;
-  }
-
-  main {
-    flex: 1 1;
-    width: 100%;
-    overflow: hidden;
-    position: relative;
   }
 </style>