+++ src/components/NavLinks.svelte
@@ -6,3 +6,4 @@
 <Link to="/about" class="link">About</Link>
 <Link to="/contact" class="link">Contact</Link>
 <Link to="/blog" class="link">Blog</Link>
+<Link to="/fullpage" class="link">Full Page</Link>