+++ src/routes/Archive.svelte
@@ -1,8 +1,0 @@
-<script>
-  import Page from "../components/Page.svelte"
-  export let location
-</script>
-
-<Page vAlign="flex-start" hAlign="center" heading="Archive" {location}>
-  <p>Old Shit</p>
-</Page>