+++ src/routes/FourOhFour.svelte
@@ -1,5 +1,5 @@
 <script>
-  import Page from "../components/Page.svelte"
+  import Page from "../components/Page2.svelte"
   export let location
 </script>