+++ src/routes/Home.svelte
@@ -1,8 +1,0 @@
-<script>
-  import Page from "../components/Page.svelte"
-  export let location
-</script>
-
-<Page vAlign="center" hAlign="center" heading="Welcome" {location}>
-  Thanks for visiting.
-</Page>