+++ src/routes/Posts.svelte
@@ -1,8 +1,0 @@
-<script>
-  import Page from "../components/Page2.svelte"
-  export let location
-</script>
-
-<Page vAlign="flex-start" hAlign="center" title="Posts" {location}>
-  <p>Posts</p>
-</Page>